Hide Patient info in batches listing when current context is Patient
